tools/requirements.txt
python-mbedtls
sslcrypto
cryptography
pycryptodome
ecdsa
pyelftools
pyserial
pyDes
